requirements.txt
django-oauth-toolkit
django-cors-headers
requests-oauthlib
django-oauth-usp
